Memory optimization re: images
In the SpriteObject, the img variable caching the image results in that entire image being held in memory when it doesn't need to be.

Replaced it with texture_diffuse_bytes and texture_normal_bytes for storing the texture data. This data is loaded in when we open a file and used to create the texture, and is what is written back out to disk when the user saves.

Replaced the anim_texture and anim_texture_normal objects given those were filling the same purpose from what I could telll, looking at the code (and those don't look to be fully implemented anyhow looking at how there's code commented out)